backend/requirements.txt
asgiref==3.8.1
dataclasses==0.6
# Django==5.0.6
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
djongo==1.2.31
dnspython==2.6.1
pymongo==4.8.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
django-oauth-toolkit
